NewsPaper/requirements.txt
amqp==5.0.9
APScheduler==3.8.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0
django-allauth==0.47.0
django-filter==21.1
idna==3.3
kombu==5.2.2
oauthlib==3.1.1
prompt-toolkit==3.0.24
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
